archunit 테스트와 패키지 구조 변경
