+++ src/main/java/com/flodiback/application/meeting/command/CreateMeetingCommand.java
@@ -1,0 +1,3 @@
+package com.flodiback.application.meeting.command;
+
+public record CreateMeetingCommand(Long projectId, String title) {}
