--- src/main/java/com/flodiback/domain/decision/decision/entity/Decision.java
@@ -1,53 +1,0 @@
-package com.flodiback.domain.decision.decision.entity;
-
-import java.time.LocalDateTime;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import com.flodiback.domain.meeting.meeting.entity.Meeting;
-import com.flodiback.domain.project.project.entity.Project;
-import com.pgvector.PGvector;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "decisions")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Decision {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "project_id", nullable = false)
-    private Project project;
-
-    // ON DELETE SET NULL → nullable
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "meeting_id", nullable = true)
-    private Meeting meeting;
-
-    @Column(name = "content", nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @Column(name = "embedding", columnDefinition = "vector(768)")
-    private PGvector embedding;
-
-    @CreationTimestamp
-    @Column(name = "decided_at", nullable = false, updatable = false)
-    private LocalDateTime decidedAt;
-
-    @Builder
-    public Decision(Project project, Meeting meeting, String content, PGvector embedding) {
-        this.project = project;
-        this.meeting = meeting;
-        this.content = content;
-        this.embedding = embedding;
-    }
-}
