--- src/main/java/com/flodiback/domain/meeting/meeting/dto/CreateMeetingRequest.java
@@ -1,3 +1,0 @@
-package com.flodiback.domain.meeting.meeting.dto;
-
-public record CreateMeetingRequest(Long projectId, String title) {}
