+++ src/main/java/com/flodiback/domain/meeting/repository/MeetingRepository.java
@@ -1,0 +1,12 @@
+package com.flodiback.domain.meeting.repository;
+
+import java.util.Optional;
+
+import com.flodiback.domain.meeting.entity.Meeting;
+
+public interface MeetingRepository {
+
+    Meeting save(Meeting meeting);
+
+    Optional<Meeting> findById(Long id);
+}
