+++ src/main/java/com/flodiback/domain/meeting/type/MeetingStatus.java
@@ -1,0 +1,6 @@
+package com.flodiback.domain.meeting.type;
+
+public enum MeetingStatus {
+    IN_PROGRESS,
+    COMPLETED
+}
