--- src/main/java/com/flodiback/domain/server/server/entity/ApiKey.java
@@ -1,43 +1,0 @@
-package com.flodiback.domain.server.server.entity;
-
-import java.time.LocalDateTime;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "api_keys")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApiKey {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "server_id", nullable = false)
-    private DiscordServer server;
-
-    @Column(name = "key_hash", nullable = false, length = 255)
-    private String keyHash;
-
-    @Column(name = "description", length = 100)
-    private String description;
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Builder
-    public ApiKey(DiscordServer server, String keyHash, String description) {
-        this.server = server;
-        this.keyHash = keyHash;
-        this.description = description;
-    }
-}
