--- src/main/java/com/flodiback/global/enums/MeetingStatus.java
@@ -1,6 +1,0 @@
-package com.flodiback.global.enums;
-
-public enum MeetingStatus {
-    IN_PROGRESS,
-    COMPLETED
-}
